removed state service
