+++ package.json
@@ -32,10 +32,13 @@
     "@progress/kendo-angular-inputs": "^7.1.0",
     "@progress/kendo-angular-intl": "^3.1.0",
     "@progress/kendo-angular-l10n": "^3.0.0",
+    "@progress/kendo-angular-label": "^3.0.0",
     "@progress/kendo-angular-pdf-export": "^3.0.0",
     "@progress/kendo-angular-popup": "^4.0.0",
+    "@progress/kendo-angular-treeview": "^5.1.0",
     "@progress/kendo-data-query": "^1.5.4",
     "@progress/kendo-drawing": "^1.9.3",
+    "@progress/kendo-licensing": "^1.1.3",
     "@progress/kendo-theme-default": "^4.32.0",
     "milligram": "^1.4.1",
     "rxjs": "^6.6.3",
