+++ projects/example/src/app/grid.directive.component.html
@@ -10,7 +10,7 @@
     info: true,
     type: 'numeric',
     pageSizes: true,
-    previousNex: true
+    previousNext: true
   }"
   [loading]="loading"
   [pageSize]="gridState.take"
