+++ projects/kendo-grid-state/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@lucasheight/kendo-grid-state",
-  "version": "0.1.12",
+  "version": "1.0.0",
   "description": "A helper library that implements a directive to manage grid state during session or between sessions for Progress Kendo UI for Angular Grid.",
   "keywords": [
     "angular",
@@ -26,21 +26,25 @@
   "peerDependencies": {
     "@angular/common": ">=8 < 10",
     "@angular/core": ">=8 < 10",
-    "@progress/kendo-angular-buttons": "5.x",
-    "@progress/kendo-angular-common": "1.x",
-    "@progress/kendo-angular-dateinputs": "4.x",
-    "@progress/kendo-angular-dropdowns": "4.x",
-    "@progress/kendo-angular-excel-export": "3.x",
-    "@progress/kendo-angular-grid": "4.x",
-    "@progress/kendo-angular-inputs": "6.x",
-    "@progress/kendo-angular-intl": "2.x",
-    "@progress/kendo-angular-l10n": "2.x",
-    "@progress/kendo-angular-pdf-export": "2.x",
-    "@progress/kendo-angular-popup": "3.x",
-    "@progress/kendo-data-query": "1.x",
-    "@progress/kendo-drawing": "1.x"
+    "@progress/kendo-angular-buttons": ">=5 < 7",
+    "@progress/kendo-angular-common": ">= 1 < 3",
+    "@progress/kendo-angular-dateinputs": ">=4 < 6",
+    "@progress/kendo-angular-dropdowns": ">=4 < 6",
+    "@progress/kendo-angular-excel-export": ">=3 < 5",
+    "@progress/kendo-angular-grid": ">= 4 < 6",
+    "@progress/kendo-angular-inputs": ">= 6 < 8",
+    "@progress/kendo-angular-intl": ">=2 < 4",
+    "@progress/kendo-angular-l10n": ">=2 < 4",
+    "@progress/kendo-angular-label": ">=2 < 4",
+    "@progress/kendo-angular-pdf-export": ">= 2 < 4",
+    "@progress/kendo-angular-popup": ">=3 < 5",
+    "@progress/kendo-angular-treeview": ">=4 < 6",
+    "@progress/kendo-data-query": ">= 1 < 2",
+    "@progress/kendo-licensing": ">= 1 < 2",
+    "@progress/kendo-theme-default": ">= 3 < 5",
+    "@progress/kendo-drawing": ">=1 <2"
   },
   "dependencies": {
-    "tslib": "^2.0.0"
+    "tslib": ">= 2 < 3"
   }
 }
