+++ projects/kendo-grid-state/src/lib/Column.ts
@@ -1,0 +1,10 @@
+export class Column {
+  origIdx?: number = 0;
+  orderIndex?: number;
+  leafIndex?: number;
+  hidden? = false;
+  width?: number;
+  expanded?: boolean = false;
+  title?: string;
+  field?: string = undefined;
+}
