+++ projects/kendo-grid-state/src/lib/GridState.ts
@@ -1,0 +1,9 @@
+import { State } from "@progress/kendo-data-query";
+import { Column } from "./Column";
+
+export interface IGridState {
+  columns: Column[];
+  state: State;
+  //  gridData?: DataResult;
+  expandedRows?: boolean[];
+}
