--- projects/kendo-grid-state/src/lib/GridStateFunctions.ts
@@ -1,13 +1,0 @@
-import { ColumnSettings } from "./ColumnSettings";
-
-export function RemoveColumnByField(
-  fieldName: string,
-  columns: ColumnSettings[]
-): ColumnSettings[] {
-  const items = [...columns];
-  items.splice(
-    items.findIndex((f) => f.field === fieldName),
-    1
-  );
-  return items;
-}
