+++ projects/kendo-grid-state/src/public-api.ts
@@ -2,9 +2,7 @@
  * Public API Surface of kendo-grid-state
  */
 
-export * from "./lib/ColumnSettings";
-export * from "./lib/GridSettings";
-export * from "./lib/GridStateFunctions";
-export * from "./lib/GridStateService";
+export * from "./lib/Column";
+export * from "./lib/GridState";
 export * from "./lib/GridStateModule";
 export * from "./lib/GridStateDirective";
